Pull creating the worker ack channel into a separate module
